requirements.txt
faster-whisper>=1.0.0
sounddevice>=0.5.0
numpy>=1.24.0
ollama>=0.4.0
pynput>=1.7.6
pyobjc-core>=10.0
pyobjc-framework-Cocoa>=10.0
pyobjc-framework-Quartz>=10.0
PyYAML>=6.0
